+++ pulse2percept/models/retina/base.py
@@ -1,4 +1,6 @@
 """:py:class:`~pulse2percept.models.retina.RetinalSpatial`"""
+import warnings
+
 import numpy as np
 
 from ..base import SpatialModel, _length_valued, _placed_coords
@@ -13,6 +15,22 @@ def _visual_field_map_first(params):
     return {'visual_field_map': params['visual_field_map'],
             **{key: val for key, val in params.items()
                if key != 'visual_field_map'}}
+
+
+def _warn_ignores_z(model, electrode_array):
+    """Warn when a model ignores nonzero electrode ``z`` coordinates.
+
+    Reads placed coordinates, so ``implant_depth`` counts as depth.
+    """
+    if np.allclose(_placed_coords(model, electrode_array,
+                                  model.space_unit)[:, 2], 0):
+        return
+    warnings.warn(
+        f"{type(model).__name__} does not model electrode-retina distance: "
+        f"nonzero z values do not change its response. In a real implant, "
+        f"distance is expected to affect stimulation threshold and spatial "
+        f"recruitment, but that relationship is not parameterized by this "
+        f"model.")
 
 
 class RetinalSpatial(SpatialModel):
